requirements.txt
python-dotenv==0.9.1
python-telegram-bot==13.7
requests==2.22.0
urllib3==1.25.8